requirements.txt
auto_mix_prep==0.2.0
decord==0.6.0
faiss_gpu==1.7.2
ftfy==6.1.3
icecream==2.1.3
numpy==1.23.1
openai==0.27.7
opencv_python==4.5.4.58
pandas==1.5.2
Pillow==9.3.0
Pillow==10.2.0
pyfunctional==1.4.3
regex==2022.7.25
SceneGraphParser==0.1.0
scikit_learn==1.1.2
scipy==1.12.0
skimage==0.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.64.1
transformers==4.28.1
torch==2.0.1+cu117
torchvision==0.15.2+cu117
